Extract weeklyIcal to its own module
